+++ swift/llm/tuner.py
@@ -2,6 +2,7 @@
 import os
 import types
 
+import numpy as np
 import torch
 import transformers
 from packaging import version
@@ -211,6 +212,62 @@ def prepare_model(model, args: SftArguments):
             optim_per_parameter=args.galore_optim_per_parameter,
         )
 
+    callbacks = []
+    if args.lisa_activated_layers > 0:
+
+        class DynamicLayerActivationCallback(TrainerCallback):
+            def __init__(self, n_layers: int, step_interval: int, model: torch.nn.Module):
+                super().__init__()
+                self.n_layers = n_layers
+                self.step_interval = step_interval
+                self.model = model
+                layers_name = None
+                layers = None
+                for name, module in model.named_modules():
+                    if isinstance(module, torch.nn.ModuleList):
+                        layers_name = name
+                        layers = module
+                        break
+                assert layers_name is not None
+                self.layers_attribute = layers_name
+                self.total_layers = len(layers)
+
+                # Freeze all layers upon initialization
+                self.freeze_all_layers()
+                self.active_layers_indices = []
+
+            def freeze_all_layers(self):
+                layers = self.model.get_submodule(self.layers_attribute)
+                for layer in layers:
+                    for param in layer.parameters():
+                        param.requires_grad = False
+
+            def on_step_begin(self, args, state, control, **kwargs):
+                # Check if it's time to switch active layers, including at step 0
+                if state.global_step % self.step_interval == 0 or state.global_step == 1:
+                    self.switch_active_layers()
+
+            def switch_active_layers(self):
+                # First, disable gradients for all layers
+                self.freeze_all_layers()
+
+                # Randomly select n_layers to activate
+                layers = eval('self.' + self.layers_attribute)  # Re-fetch layer references
+                self.active_layers_indices = np.random.choice(range(self.total_layers), self.n_layers,
+                                                              replace=False)
+                print(f"Activating layers at indices: {self.active_layers_indices} for the next steps.")
+
+                # Enable gradients only for the selected layers
+                for idx in self.active_layers_indices:
+                    for param in layers[idx].parameters():
+                        param.requires_grad = True
+
+        callbacks.append(DynamicLayerActivationCallback(
+            n_layers=args.lisa_activated_layers,  # Number of layers to activate
+            step_interval=args.lisa_step_interval,  # Step interval to update active layers
+            model=model)
+        )
+
     class TrainerAdapterCallback(TrainerCallback):
 
         def __init__(self):
@@ -234,7 +291,6 @@ def prepare_model(model, args: SftArguments):
             if args.sft_type == 'adalora':
                 self.global_step = state.global_step
 
-    callbacks = []
     if is_adapter(args.sft_type) and args.tuner_backend == 'swift':
         callbacks.append(TrainerAdapterCallback())
     return model, callbacks
